refactor: moves cel to common crate
